Using latest version of org-mode
The 9.4.4 version that comes with Emacs conflicts with a newer version
I install. This is because I need org before I tangle my files.

So, I've moved the straight and `(use-package org)` code to the
`init.el` file, and then the rest of the system comes right up with
the latest org, and without the conflicts.