Swapped precedence from snipe to surround
Since I often use avy to do my jumping, snipe isn't as important, so I
limit it to just normal and visual modes, and let 's' be the surround
key, as I seem to want to use that a lot.

Also, I finally learned that `o` is the symbol in inner/outer text
objects, so that takes care of that use case.