Send code blocks from Markdown to current Terminal
Lot of terminal changes here ...

  * Single command to split window and launch terminal
  * Better terminal names based on the project
  * Send current line (or code block) to project's terminal

This last bit is attempting to dwim based on major mode and context.
So far, it is nice to read a README.md, and send the commands my
teammates write to a terminal, almost as if I were executing the
commands in an org-mode file.