Label the `g` commands for which-key
Cleaned up the commands that follow on the g key with labels, and I
got rid of commands that I don't think is useful.

This required changes through the repo, but things are looking good.

I also changed the behavior of `e` and `E`, and `w` and `W` based on
how I think that should be.