Migrating my local leader key to major-mode-hydra
Seems that the leader has a bug when it comes to local mode maps, and
the major-mode hydra is an attractive alternative. Granted, I do feel
the need to group things for the tabular view, but that isn't a bad thing.

Need to take a pass at my more-complex code in my programming stuff.