Update dslide-based demonstrations
I'm really digging the dslide approach, for most things, this is
easier to debug and work with. You still need to "mostly" be in the
presentation to get it to work.

I might have to have F5 pop to the presentation buffer, and then issue
a dslide next slide command ...